backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
cryptography==43.0.1
pycryptodome==3.23.0
